requirements.txt
fastapi==0.100.0
python-dotenv==1.0.0
python-jose==3.3.0
SQLAlchemy==2.0.19
